// src/services/generated/models/ProgramResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Fitlytics API
 * Workout tracking and training analytics API. Protected routes require a WorkOS access token in the `Authorization: Bearer ...` header.
 *
 * The version of the OpenAPI document: 0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ProgramWeekResponse } from './ProgramWeekResponse';
import {
    ProgramWeekResponseFromJSON,
    ProgramWeekResponseFromJSONTyped,
    ProgramWeekResponseToJSON,
    ProgramWeekResponseToJSONTyped,
} from './ProgramWeekResponse';

/**
 * 
 * @export
 * @interface ProgramResponse
 */
export interface ProgramResponse {
    /**
     * 
     * @type {string}
     * @memberof ProgramResponse
     */
    createdAt?: string;
    /**
     * 
     * @type {string}
     * @memberof ProgramResponse
     */
    description?: string;
    /**
     * 
     * @type {string}
     * @memberof ProgramResponse
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof ProgramResponse
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof ProgramResponse
     */
    updatedAt?: string;
    /**
     * 
     * @type {Array<ProgramWeekResponse>}
     * @memberof ProgramResponse
     */
    weeks?: Array<ProgramWeekResponse>;
}

/**
 * Check if a given object implements the ProgramResponse interface.
 */
export function instanceOfProgramResponse(value: object): value is ProgramResponse {
    return true;
}

export function ProgramResponseFromJSON(json: any): ProgramResponse {
    return ProgramResponseFromJSONTyped(json, false);
}

export function ProgramResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): ProgramResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'createdAt': json['created_at'] == null ? undefined : json['created_at'],
        'description': json['description'] == null ? undefined : json['description'],
        'id': json['id'] == null ? undefined : json['id'],
        'name': json['name'] == null ? undefined : json['name'],
        'updatedAt': json['updated_at'] == null ? undefined : json['updated_at'],
        'weeks': json['weeks'] == null ? undefined : ((json['weeks'] as Array<any>).map(ProgramWeekResponseFromJSON)),
    };
}

export function ProgramResponseToJSON(json: any): ProgramResponse {
    return ProgramResponseToJSONTyped(json, false);
}

export function ProgramResponseToJSONTyped(value?: ProgramResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'created_at': value['createdAt'],
        'description': value['description'],
        'id': value['id'],
        'name': value['name'],
        'updated_at': value['updatedAt'],
        'weeks': value['weeks'] == null ? undefined : ((value['weeks'] as Array<any>).map(ProgramWeekResponseToJSON)),
    };
}